Various robustness fixes.
Rev: src/modules/_WhiteFish/blob.c:1.33
Rev: src/modules/_WhiteFish/blob.h:1.15
Rev: src/modules/_WhiteFish/blobs.c:1.11
Rev: src/modules/_WhiteFish/buffer.c:1.12
Rev: src/modules/_WhiteFish/linkfarm.c:1.4
Rev: src/modules/_WhiteFish/resultset.c:1.22

+++ src/modules/_WhiteFish/blob.h
@@ -66,7 +66,7 @@ int wf_blob_eof( Blob *b );
 void wf_blob_low_add( struct object *o, int docid, int field, int off );
 /* Add a hit */
 
-int wf_blob_low_memsize( struct object *o );
+size_t wf_blob_low_memsize( struct object *o );
 /* Returns the size of the blob */
 
 void init_blob_program();
